==> step8-prototype.ipynb <==
# Prototype for Twitter Sentiment Analysis for Zoom Stock

This Prototype will implement the skeleton of the final project.

From Step 7, this step will use SVM and RoBERTA (non-fine tuned) to analyze 5000 zoom stock tweets.


using labeled tweets to train SVM and sentiment analysis  of zoom tweets


# load tweets. note that final project will have all tweets that have sentiment analysis done.

import pandas as pd




# zm_tweet_df = pd.read_csv('zm-tweet-2020-2-26-to-2020-2-29.csv')
zm_tweet_df = pd.read_csv('zm-tweets-jan to april 2020.csv')


# print(zm_tweet_df.head())

print(zm_tweet_df[['date', 'cashtags']].head())

# print(zm_tweet_df['date'].tail())
print(zm_tweet_df[['date', 'cashtags']].tail())

print(zm_tweet_df.shape)

# Training SVM

# loading labelled sentiment twitter posts


# import pandas as pd
# import re


import numpy as np
import pandas as pd
import nltk
from nltk.corpus import stopwords
import matplotlib.pyplot as plt
from nltk.stem import WordNetLemmatizer
from sklearn.feature_extraction.text import CountVectorizer
import pickle
import time
import re

labeled_tweet_df = pd.read_csv("dataset-labled-tweets-original.csv")

# cleaning tweets for sentiment analysis
import re

def handle_emojis(tweet):
    # Smile -- :), : ), :-), (:, ( :, (-:, :')
    tweet = re.sub(r'(:\s?\)|:-\)|\(\s?:|\(-:|:\'\))', ' EMO_POS ', tweet)
    # Laugh -- :D, : D, :-D, xD, x-D, XD, X-D
    tweet = re.sub(r'(:\s?D|:-D|x-?D|X-?D)', ' EMO_POS ', tweet)
    # Love -- <3, :*
    tweet = re.sub(r'(<3|:\*)', ' EMO_POS ', tweet)
    # Wink -- ;-), ;), ;-D, ;D, (;,  (-;
    tweet = re.sub(r'(;-?\)|;-?D|\(-?;)|😉', ' EMO_POS ', tweet)
    # Sad -- :-(, : (, :(, ):, )-:
    tweet = re.sub(r'(:\s?\(|:-\(|\)\s?:|\)-:)', ' EMO_NEG ', tweet)
    # Cry -- :,(, :'(, :"(
    tweet = re.sub(r'(:,\(|:\'\(|:"\()', ' EMO_NEG ', tweet)

    return tweet

def remove_emoji(string):
    emoji_pattern = re.compile("["
                           u"\U0001F600-\U0001F64F" # emoticons
                           u"\U0001F300-\U0001F5FF" # symbols & pictographs
                           u"\U0001F680-\U0001F6FF" # transport & map symbols
                           u"\U0001F1E0-\U0001F1FF" # flags (iOS)
                           u"\U00002500-\U00002BEF"  # chinese char
                           u"\U00002702-\U000027B0"
                           u"\U00002702-\U000027B0"
                           u"\U000024C2-\U0001F251"
                           u"\U0001f926-\U0001f937"
                           u"\U00010000-\U0010ffff"
                           u"\u2640-\u2642"
                           u"\u2600-\u2B55"
                           u"\u200d"
                           u"\u23cf"
                           u"\u23e9"
                           u"\u231a"
                           u"\ufe0f"  # dingbats
                           u"\u3030"
                           "]+", flags=re.UNICODE)
    
    return emoji_pattern.sub(r'', string)


import nltk
from nltk.corpus import stopwords
from nltk.stem import PorterStemmer, WordNetLemmatizer

def preprocess_word(word):
    # Tokenize the word
    tokens = nltk.word_tokenize(word)
    
    # Convert to lowercase
    tokens = [token.lower() for token in tokens]
    
    # Remove stopwords
    stop_words = set(stopwords.words('english'))
    tokens = [token for token in tokens if token not in stop_words]
    
    # Stem the tokens
    stemmer = PorterStemmer()
    tokens = [stemmer.stem(token) for token in tokens]
    
    # Lemmatize the tokens
    lemmatizer = WordNetLemmatizer()
    tokens = [lemmatizer.lemmatize(token) for token in tokens]
    
    # Join the tokens back into a single string
    preprocessed_word = ' '.join(tokens)
    
    return preprocessed_word



def preprocess_tweet(tweet):
    processed_tweet = []
    # Convert to lower case
    tweet = tweet.lower()
    
    #Clean only digits
    tweet = re.sub("^\d+\s|\s\d+\s|\s\d+$", " ", tweet)
    
    # Replaces URLs with the word URL
    #tweet = re.sub(r'((www\.[\S]+)|(https?://[\S]+))', ' URL ', tweet)
    tweet = re.sub(r'((www\.[\S]+)|(https?://[\S]+))', '', tweet)
    
    # Replace @handle with the word USER_MENTION
    #tweet = re.sub(r'@[\S]+', 'USER_MENTION', tweet)
    tweet = re.sub(r'@[\S]+', '', tweet)
    
    # Replaces #hashtag with hashtag
    #tweet = re.sub(r'#(\S+)', r' \1 ', tweet)
    tweet = re.sub(r'#(\S+)', '', tweet)
    
    # Remove RT (retweet)
    tweet = re.sub(r'\brt\b', '', tweet)
    
    # Replace 2+ dots with space
    tweet = re.sub(r'\.{2,}', ' ', tweet)
    
    # Strip space, " and ' from tweet
    tweet = tweet.strip(' "\'')

    # Replace emojis with either EMO_POS or EMO_NEG
    #tweet = handle_emojis(tweet)
    tweet = remove_emoji(tweet)
   
    # Replace multiple spaces with a single space
    tweet = re.sub(r'\s+', ' ', tweet)

    #my custom chars
    tweet = tweet.replace('₺','')
    tweet = tweet.replace('=','')
    tweet = tweet.replace('’','')
    tweet = tweet.replace('|','')
    tweet = tweet.replace('‘','')
    tweet = tweet.replace('/','')
    tweet = tweet.replace('…','')
    tweet = tweet.replace('–','')
    tweet = tweet.replace('&','')
    tweet = tweet.replace('“','')
    tweet = tweet.replace('”','')
    tweet = tweet.replace('+','')
    tweet = tweet.replace('%','')
    tweet = tweet.replace('@','')
    tweet = tweet.replace('#','')

    words = word_tokenize(tweet) #tweet.split()

    for word in words:
      word = preprocess_word(word)
      #if is_valid_word(word):
      #    processed_tweet.append(word)
      processed_tweet.append(word)

    return ' '.join(processed_tweet)


from nltk import word_tokenize

import nltk
nltk.download('punkt')
nltk.download('wordnet')
nltk.download('omw-1.4')

def clean_tweet(tweet):
    tweet = str(tweet)
    handle_emojis(tweet)
    remove_emoji(tweet)
    preprocess_tweet(tweet)
    return str(tweet) 




labeled_tweet_df["clean_tweet"] = labeled_tweet_df['text'].apply(clean_tweet) # applies clean_tweet function on each record of 'text' and creates a new column 'clean_tweet'

labeled_tweet_df.head()

labeled_tweet_df.shape

from sklearn.model_selection import train_test_split

train, test = train_test_split(labeled_tweet_df, test_size=0.2)
test.head


vectorizer = CountVectorizer(max_features=10000)
BOW = vectorizer.fit_transform(labeled_tweet_df['clean_tweet'])


from sklearn.model_selection import train_test_split
x_train,x_test,y_train,y_test = train_test_split(BOW,np.asarray(labeled_tweet_df["sentiment"]))




from sklearn.svm import SVC
model = SVC()
model.fit(x_train,y_train)


predictions = model.predict(x_test)

print(type(predictions))

print(len(predictions))

from sklearn.metrics import accuracy_score,confusion_matrix

print("Accuracy of model is {}%".format(accuracy_score(y_test,predictions) * 100))


# note that accuracy changes each time: so far ranging from 67.37% to 69%

# Clean Zoom Tweets



# cleaning zoom tweets using the following method
# df["clean_tweet"] = df['text'].apply(clean_tweet) # applies clean_tweet function on each record of 'text' and creates a new column 'clean_tweet'

zm_tweet_df['clean_tweet'] = zm_tweet_df['content'].apply(clean_tweet)



zm_tweet_df[['clean_tweet', 'cashtags']].tail()

# Using Trained SVM Model for sentiment analysis for 5k Zoom Twitter Dataset

x_test

# run svm on 5k dataset 
    # create a new column "svm_zm_sentiment" on zm_tweet_df['clean_tweet']
        #setup SVM for sentiment analysis - setting up Bag of Words (BOW) - note that this was all done earlier
new_BOW = vectorizer.transform(zm_tweet_df['clean_tweet'])

        #create the ndarray predictions_svm_zm to be added as a new column back into zm_tweet_df
predictions_svm_zm = model.predict(new_BOW)
        #creating the new column 'svm_sentiment'
zm_tweet_df['sentiment_svm'] = predictions_svm_zm


# about 60 sec for SVM for 5k tweets note how much time passed. to compare to RoBERTa sentiment analysis





# quick look at svm_sentiment

zm_tweet_df[['date', 'clean_tweet', 'sentiment_svm']].tail()

zm_tweet_df['sentiment_svm'].value_counts()


# Save to entire df to csv 'zm_5k_tweets_step1_with_svm.csv'
zm_tweet_df.to_csv('zm_5k_tweets_step1_with_svm.csv')



# RoBERTa - no finetuning twitter analysis on 5k twitter dataset




# code to load RoBERTa from step 7 (this step, no time to use CUDA cores will use old slow for loop method)
# load zm_5k_tweets_step1_with_svm.csv into zm_tweet_df (run this if not running from start of notebook)


# uncomment if running from start
zm_tweet_df = pd.read_csv('zm_5k_tweets_step1_with_svm.csv') 


# Prep for RoBERTa 

from transformers import AutoModelForSequenceClassification
from transformers import TFAutoModelForSequenceClassification
from transformers import AutoTokenizer, AutoConfig
import numpy as np
from scipy.special import softmax
import csv
import urllib.request
import re

task='sentiment'
MODEL = f"cardiffnlp/twitter-roberta-base-{task}"

print(f"\n\nMODEL string: {MODEL}\n\n")
tokenizer = AutoTokenizer.from_pretrained(MODEL)

# download label mapping
labels=[]
mapping_link = f"https://raw.githubusercontent.com/cardiffnlp/tweeteval/main/datasets/{task}/mapping.txt"
with urllib.request.urlopen(mapping_link) as f:
    html = f.read().decode('utf-8').split("\n")
    csvreader = csv.reader(html, delimiter='\t')
labels = [row[1] for row in csvreader if len(row) > 1]

model = AutoModelForSequenceClassification.from_pretrained(MODEL)

# Testing individual tweets

text = "$zm ended the day at the price of $90.95"
text = clean_tweet(text)
encoded_input = tokenizer(text, return_tensors='pt')
output = model(**encoded_input)
scores = output[0][0].detach().numpy()
scores = softmax(scores)

print(f"Results on text: {text}")


ranking = np.argsort(scores)

print("ranking: " + str(ranking))
ranking = ranking[::-1]
print("ranking: " + str(ranking))
for i in range(scores.shape[0]):
    l = labels[ranking[i]]
    s = scores[ranking[i]]
    print(f"{i+1}) {l} {np.round(float(s), 4)}")


MODEL_R = f"cardiffnlp/twitter-roberta-base-sentiment-latest"
tokenizer = AutoTokenizer.from_pretrained(MODEL)
config = AutoConfig.from_pretrained(MODEL)
model_r = AutoModelForSequenceClassification.from_pretrained(MODEL)

zm_tweet_df['clean_tweet'].tail()


zm_tweet_df['sentiment_svm'].tail(30)

# windows - no sleep  function

# RoBERTa predicted sentiments
predictions = []

for text in zm_tweet_df['clean_tweet']:
    encoded_input = tokenizer(text, return_tensors='pt')
    output = model_r(**encoded_input)
    scores = output[0][0].detach().numpy()
    scores = softmax(scores)

    ranking = np.argsort(scores)
    ranking = ranking[::-1]
    predicted_label = config.id2label[ranking[0]]
    predictions.append(predicted_label)

# 540m 52.6s - time taken for 5683 RoBERTa analysis
# windows - stop no sleep function (operation done and computer can auto sleep)


# print display after each 50 processed.. print row_num + seconds_passed
# when complete. note how much time passed. compare to SVM time.

# 540m 52.6s - time taken for 5683 RoBERTa analysis
# compared to about 60 sec for SVM for 5k tweets note how much time passed. to compare to RoBERTa sentiment analysis

# prep to add new col 'r_sentiment'
df_predictions_r_series = pd.Series(predictions,  name='sentiment_r') # creating temp df series 

# create a new column in 5k zoom tweet 'r_sentiment'
zm_tweet_df = zm_tweet_df.assign(sentiment_r = df_predictions_r_series) # creating the new col


# checking to see if all the sentiment cols are there
zm_tweet_df[['clean_tweet', 'sentiment_svm', 'sentiment_r', 'cashtags']].head(40)

# Labels: 0 -> Negative; 1 -> Neutral; 2 -> Positive

# converting roberta label to text

zm_tweet_df['sentiment_r_text'] = zm_tweet_df['sentiment_r'].apply(lambda x: 'neutral' if x == 'LABEL_1' else ('negative' if x == 'LABEL_0' else 'positive'))

zm_tweet_df[['clean_tweet', 'sentiment_svm', 'sentiment_r', 'sentiment_r_text', 'cashtags']].head(40)


# save entire df with 'zm_5k_tweets_with_svm_and_r.csv'
zm_tweet_df.to_csv('zm_5k_tweets_step2_with_svn_and_roberta.csv')


# Compare differences in sentiment analysis between SVM and RoBERTa

This will help me formulate the best way to finetune RoBERTa to be more accurate.

If there is time, we should finetune using the labeled tweets too. 


import pandas as pd
# load df from 'zm_5k_tweets_step2_with_svn_and_roberta.csv'
zm_tweet_df = pd.read_csv('zm_5k_tweets_step2_with_svn_and_roberta.csv')



# quick check - svm -- how many positive, neutral and negative using the 1 line code
print('sentiment_svm counts')
print(zm_tweet_df['sentiment_svm'].value_counts() )
    # create graph for better visualization
# quick check - roberta -- how many positive, neutral and negative
print('sentiment_r_text counts')
print(zm_tweet_df['sentiment_r_text'].value_counts())
    # create graph for better visualization

# quite a big difference in counts.
# sentiment_svm counts
# neutral     5094
# positive     423
# negative     167
# Name: sentiment_svm, dtype: int64
# sentiment_r_text counts
# neutral     3687
# positive    1106
# negative     891
# Name: sentiment_r_text, dtype: int64





# quick check - how many svm and roberta MATCHES
    # loop through each row and count matches or mismatch - save into a new column zm_tweet_df['match']
zm_tweet_df['match'] = zm_tweet_df['sentiment_svm'] == zm_tweet_df['sentiment_r_text']


    # of the matches , show how many postive, negative and neutral

print(len(zm_tweet_df))
print('Matches')
print(zm_tweet_df['match'].value_counts())

# out of 5684 tweets, matches 3993 and non-match 1691 ... so I have to go through 1691 tweets
# True     3993
# False    1691
# Name: match, dtype: int64




# filter out all the cashtags with more than 3

# zm_mismatch_df = zm_tweet_df[(zm_tweet_df['match']==False) & (len(zm_tweet_df['cashtags']) <3)]
# zm_tweet_df['len_cashtags'] = zm_tweet_df['cashtags'].apply(len)
cashtags_value_count_list = zm_tweet_df['cashtags'].value_counts()

print(cashtags_value_count_list.head(40))

# len(zm_tweet_df.loc[0, 'cashtags'])



# Print the first 10 rows of the 'cashtags' column and their types
for idx, value in enumerate(zm_tweet_df['cashtags'].head(10)):
    print(f"Row {idx}: {value} - Type: {type(value)}")



import pandas as pd
import ast

# Replace zm_tweet_df with your actual DataFrame
zm_tweet_df['cashtags'] = zm_tweet_df['cashtags'].apply(lambda x: ast.literal_eval(x) if isinstance(x, str) else x)

# Now, get the length of each list in the 'cashtags' column
lengths = zm_tweet_df['cashtags'].apply(lambda x: len(x) if isinstance(x, list) else 0)
print(lengths)

zm_tweet_df['cashtags_len'] = lengths


# # Print the first 10 rows of the 'cashtags' column and their types
# for idx, value in enumerate(zm_tweet_df['cashtags_len'].head(10)):
#     print(f"Row {idx}: {value} - Type: {type(value)}")

zm_tweet_filtered_cashtags_3 = zm_tweet_df[ zm_tweet_df['cashtags_len'] <= 3]
len(zm_tweet_filtered_cashtags_3)

pd.set_option('display.max_colwidth', 550)  # Set a specific value


zm_mismatch_df = zm_tweet_filtered_cashtags_3[zm_tweet_filtered_cashtags_3['sentiment_svm'] != zm_tweet_filtered_cashtags_3['sentiment_r_text']]
print(len(zm_mismatch_df))
# zm_mismatch_df['content', 'sentiment_svm', 'sentiment_r_text'].head(20)
# zm_mismatch_df[['content', 'sentiment_svm', 'sentiment_r_text', 'cashtags']].head(40)
zm_mismatch_df.iloc[130:200][['content', 'sentiment_svm', 'sentiment_r_text', 'cashtags']].head(40)
# zm_mismatch_df.columns


# Save zm_mismatch_df to csv

zm_mismatch_df.to_csv('zm_mismatch_df.csv')


# quick check - how  many svm and roberta DOESN'T Match --- ans: # False    1691
    # of the mismatch, show how many positive, negative and neutral (this will show me how far off both models are)

print(zm_tweet_df.loc[zm_tweet_df['match'] == False, ['content', 'sentiment_svm', 'sentiment_r_text', 'cashtags']])
